README notes on overriding system prompt, plus improvements & clarity to system prompt prep code. Fix server startup.
